Split api into seperate pagemaker
